feat: validate config
