+++ packages/cli-tools/src/client.ts
@@ -1,43 +1,11 @@
-import path from 'path'
-import os from 'os'
-import { readFileSync } from 'fs'
 import _ from 'lodash'
 import { BrubeckClientConfig, StreamrClient, ConfigTest } from 'streamr-client'
 import { GlobalCommandLineArgs } from './common'
+import { getConfig } from './config'
 
-const tryReadFile = (fileName: string): string|undefined => {
-    try {
-        return readFileSync(fileName, 'utf8')
-    } catch (e: any) {
-        return undefined
-    }
-}
-
-const getConfigFileJson = (id?: string): any|undefined => {
-    const CONFIG_DIRECTORY = path.join(os.homedir(), '.streamr', 'config')
-    let fileNames: string[]
-    if (id !== undefined) {
-        fileNames = [
-            id,
-            `${id}.json`,
-            path.join(CONFIG_DIRECTORY, `${id}.json`),
-            
-        ]
-    } else {
-        fileNames = [ path.join(CONFIG_DIRECTORY, `default.json`) ]
-    }
-    for (const fileName of fileNames) {
-        const content = tryReadFile(fileName)
-        if (content !== undefined) {
-            return JSON.parse(content).client
-        }
-    }
-    return undefined
-}
-
-const getConfig = (commandLineArgs: GlobalCommandLineArgs, overridenOptions: BrubeckClientConfig) => {
+const getClientConfig = (commandLineArgs: GlobalCommandLineArgs, overridenOptions: BrubeckClientConfig) => {
     const environmentOptions = (commandLineArgs.dev !== undefined) ? _.omit(ConfigTest, 'auth') : undefined
-    const configFileJson = getConfigFileJson(commandLineArgs.config)
+    const configFileJson = getConfig(commandLineArgs.config)?.client
     const authenticationOptions = (commandLineArgs.privateKey !== undefined) ? { auth: { privateKey: commandLineArgs.privateKey } } : undefined
     return _.merge(
         environmentOptions,
@@ -48,6 +16,6 @@ const getConfig = (commandLineArgs: GlobalCommandLineArgs, overridenOptions: Bru
 }
 
 export const createClient = (commandLineArgs: GlobalCommandLineArgs, overridenOptions: BrubeckClientConfig = {}): StreamrClient => {
-    const config = getConfig(commandLineArgs, overridenOptions)
+    const config = getClientConfig(commandLineArgs, overridenOptions)
     return new StreamrClient(config)
 }
